# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Asesora-T API Doc",
        "description": "Documentación de la API de Asesora-T",
        "contact": {
            "email": "caamalmarco99@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://asesorat-backend.test:8080",
            "description": "SwaggerHUB API Mocking"
        }
    ],
    "paths": {
        "/api/cuentas/login": {
            "post": {
                "tags": [
                    "Cuentas"
                ],
                "summary": "Autenticarse en la API",
                "description": "Aquí el usuario puede autenticarse en la API.",
                "operationId": "login",
                "requestBody": {
                    "description": "Credenciales del usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CredencialesUsuario"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CredencialesUsuario": {
                "title": "Credenciales Usuario",
                "description": "Credenciales del usuario necesarias para poder iniciar sesión.",
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}